app/requirements.txt
boto3==1.28.64
botocore==1.31.64
langchain==0.1.14
numpy==1.26.4
opensearch-py==2.5.0
pandas==2.1.4
pyyaml==6.0.1
streamlit==1.32.0
tenacity==8.2.3